expanded on game state schema, moved code around
