+++ server/src/rooms/Lobby.ts
@@ -1,14 +1,13 @@
 import { Room, Client } from "@colyseus/core";
-import { LobbyState } from "./schema/LobbyState";
-import { PlayerState } from "./schema/PlayerState";
+import { PlayerState, GameState } from "./schema/CodeRedState";
 
 // TODO: Merge lobby state and game state into one?
-export class Lobby extends Room<LobbyState> {
+export class CodeRedRoom extends Room<GameState> {
   // Allow up to 6 players per room
   maxClients = 6;
 
   onCreate(options: any) {
-    this.setState(new LobbyState());
+    this.setState(new GameState());
     this.state.roomCode = this.generateRoomCode();
 
     // Create all our event handlers down below
