--- server/src/rooms/schema/LobbyState.ts
@@ -1,8 +1,0 @@
-import { Schema, MapSchema, type } from "@colyseus/schema";
-import { PlayerState } from "./PlayerState";
-
-export class LobbyState extends Schema {
-  @type({ map: PlayerState }) players = new MapSchema<PlayerState>();
-  @type("string") roomCode: string;
-  @type("string") hostId: string;
-}
