--- server/src/rooms/schema/PlayerState.ts
@@ -1,5 +1,0 @@
-import { Schema, type } from "@colyseus/schema";
-
-export class PlayerState extends Schema {
-  @type("string") name: string;
-}
